[master] code refactorings

+++ src/playground/memory/v1/memory.c
@@ -36,7 +36,7 @@ static void memory_free(const void* data, u64 size) {
     ZEROPTR(data)
     ptr -= size;
 #ifdef USE_MEMORY_DEBUG_INFO
-    printf("   -: 0x%016llx !  %16lld\n", (u64)ptr, (u64)size);
+    printf("   -: 0x%016llx !  %16lld\n", (u64)ptr, size);
 #endif
 }
